requirements.txt
streamlit
yt-dlp
openai
openai-whisper
torch
tqdm
